requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.6
scikit-learn
mlflow==2.7.1
pyaml
dagshub
fastapi
uvicorn
python-multipart


##-e .